Restructure CLI into a package with per-integration app groups
Split src/pincer/cli.py into src/pincer/cli/ with dedicated Typer
subgroups: google (setup/auth), signal (setup/status/test), slack
(setup), and whatsapp (setup), replacing the old flat pair-whatsapp,
setup-slack, auth-google, and setup-google commands, and renaming
signal's pair task to setup for consistency. Updates all in-code hints
and documentation (CLI reference, guides, project structure, tools
catalog) to match, including removing stale references to the old
cli.py module and to CLI flags/commands that no longer exist.
