+++ docs/core-components/tools.md
@@ -6,9 +6,9 @@ This is the complete, authoritative catalog of every tool Pincer can call. It is
 
 | Category | Tools | Module | How to enable |
 |---|---:|---|---|
-| **Core built-ins** (incl. `generate_image`, `make_phone_call`, `load_skill`, `load_skill_reference`, `run_skill_script`) | 26 | `src/pincer/tools/builtin/` + `src/pincer/cli.py` | Always on (voice/image need their API key) |
+| **Core built-ins** (incl. `generate_image`, `make_phone_call`, `load_skill`, `load_skill_reference`, `run_skill_script`) | 26 | `src/pincer/tools/builtin/` + `src/pincer/cli/run.py` | Always on (voice/image need their API key) |
 | **Bundled skills** | 5 (progressive disclosure, no fixed tool count) | `src/pincer/skills/` (SKILL.md) | Always on (bundled) |
-| **Google Workspace** | 113 | `src/pincer/integrations/google/` | `pincer setup-google` |
+| **Google Workspace** | 113 | `src/pincer/integrations/google/` | `pincer google setup` |
 | **Microsoft 365** | 69 | `src/ms365-mcp/ms365/` | `ms365-mcp-setup` |
 | **Slack (native)** | 71 | `src/pincer/integrations/slack/` | `PINCER_SLACK_BOT_TOKEN` |
 | **MCP servers (external)** | unlimited | `src/pincer/mcp/` | `pincer.toml` `[[mcp.servers]]` |
@@ -22,7 +22,7 @@ This is the complete, authoritative catalog of every tool Pincer can call. It is
 
 ## Core built-ins (26)
 
-Registered in `src/pincer/cli.py::_run_agent()`. These are always available.
+Registered in `src/pincer/cli/run.py::_run_agent()`. These are always available.
 
 | Tool | Approval | Description |
 |---|:---:|---|
@@ -71,7 +71,7 @@ Discovered from `src/pincer/skills/` (shipped inside the package) at startup —
 
 ## Google Workspace (113 tools)
 
-Registered when the user completes `pincer setup-google`. All tools are prefixed `google__`.
+Registered when the user completes `pincer google setup`. All tools are prefixed `google__`.
 
 ### Gmail (19)
