+++ src/pincer/__main__.py
@@ -1,5 +1,5 @@
 """Allow running with `python -m pincer`."""
 
-from pincer.cli import app
+from pincer.cli.main import app
 
 app()
