+++ src/pincer/cli/__init__.py
@@ -1,0 +1,3 @@
+from pincer.cli.main import app
+
+__all__ = ["app"]
